+++ packages/devtools/vite.config.ts
@@ -1,18 +1,15 @@
 import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
+import react, { reactCompilerPreset } from "@vitejs/plugin-react";
 import tailwindcss from "@tailwindcss/vite";
+import babel from "@rolldown/plugin-babel"
 
 export default defineConfig(({ command }) => ({
   base: command === "build" ? "./" : "/",
   plugins: [
-    react({
-      babel: {
-        plugins: [["babel-plugin-react-compiler", { target: "19" }]],
-      },
+    react(),
+    babel({
+      presets: [reactCompilerPreset()]
     }),
     tailwindcss(),
   ],
-  server: {
-    port: 5173,
-  },
 }));