+++ samples/counter/src/env.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="@fcannizzaro/streamdeck-react/font" />