+++ samples/jotai/src/plugin.ts
@@ -1,19 +1,13 @@
-import { createPlugin } from "@fcannizzaro/streamdeck-react";
+import { createPlugin, googleFont } from "@fcannizzaro/streamdeck-react";
 import { displayAction } from "./actions/display";
 import { incrementAction } from "./actions/increment";
 import { resetAction } from "./actions/reset";
 import { JotaiWrapper } from "./wrapper";
-import InterRegular from "@fontsource-variable/inter/files/inter-latin-wght-normal.woff2";
+
+const inter = await googleFont("Inter");
 
 const plugin = createPlugin({
-  fonts: [
-    {
-      name: "Inter",
-      data: InterRegular,
-      weight: 400,
-      style: "normal",
-    },
-  ],
+  fonts: [inter],
   actions: [displayAction, incrementAction, resetAction],
   wrapper: JotaiWrapper,
 });