+++ samples/pokemon/src/plugin.ts
@@ -1,17 +1,11 @@
-import { createPlugin } from "@fcannizzaro/streamdeck-react";
+import { createPlugin, googleFont } from "@fcannizzaro/streamdeck-react";
 import { pokemonAction } from "./actions/pokemon";
 import { QueryWrapper } from "./wrapper";
-import InterRegular from "@fontsource-variable/inter/files/inter-latin-wght-normal.woff2";
+
+const inter = await googleFont("Inter");
 
 const plugin = createPlugin({
-  fonts: [
-    {
-      name: "Inter",
-      data: InterRegular,
-      weight: 400,
-      style: "normal",
-    },
-  ],
+  fonts: [inter],
   devtools: true,
   actions: [pokemonAction],
   wrapper: QueryWrapper,